feat: make conversion optional in persistence layer
