--- commons-data-impl/src/main/java/com/appjars/saturn/dao/ConversionCrudDaoMixin.java
@@ -1,30 +1,0 @@
-/*-
- * #%L
- * Commons Backend - Data Access Layer Implementations
- * %%
- * Copyright (C) 2020 Flowing Code
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package com.appjars.saturn.dao;
-
-import java.io.Serializable;
-
-import com.appjars.saturn.conversion.ConversionSupport;
-import com.appjars.saturn.model.Identifiable;
-
-public interface ConversionCrudDaoMixin<S extends Serializable, T extends Identifiable<K>, K extends Serializable>
-		extends CrudDao<S, K>, ConversionSupport<S, T> {
-
-}
