+++ commons-data-impl/src/test/java/com/appjars/saturn/dao/PersonCrudDaoImpl.java
@@ -26,7 +26,7 @@ import com.appjars.saturn.dao.jpa.JpaDaoSupport;
 import com.appjars.saturn.model.impl.Person;
 
 public class PersonCrudDaoImpl
-		implements JpaDaoSupport<Person, Person, Integer>, EntityConversionCrudDaoMixin<Person, Integer> {
+		implements JpaDaoSupport<Person, Person, Integer> {
 
 	EntityManagerFactory entityManagerFactory;
